Make Bforartists addons "Official" #4121 - Made the Xray Weight Paint addon a Bforartists one
