+++ scripts/addons/bfa_xray_weight_paint_button/__init__.py
@@ -1,0 +1,52 @@
+# ##### BEGIN GPL LICENSE BLOCK #####
+#
+#  This program is free software; you can redistribute it and/or
+#  modify it under the terms of the GNU General Public License
+#  as published by the Free Software Foundation; either version 3
+#  of the License, or (at your option) any later version.
+#
+#  This program is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU General Public License for more details.
+#
+#  You should have received a copy of the GNU General Public License
+#  along with this program; if not, write to the Free Software Foundation,
+#  Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
+#
+# ##### END GPL LICENSE BLOCK #####
+
+bl_info = {
+    "name": "X-Ray Weight Paint Button",
+    "author": "BlenderBoi",
+    "version": (1, 0),
+    "blender": (2, 80, 0),
+    "location": "3D View",
+    "description": "Add an Xray Weight Paint Button in 3D View header",
+    "warning": "",
+    "doc_url": "https://github.com/Bforartists/Manual",
+    "tracker_url": "https://github.com/Bforartists/Bforartists",
+    "category": "Bforartists",
+    "support": "OFFICIAL",
+}
+
+
+
+from . import WeightPaintBrush_Menu
+
+
+modules =  [WeightPaintBrush_Menu] 
+
+def register():
+
+
+    for module in modules:
+        module.register()
+
+def unregister():
+
+    for module in modules:
+        module.unregister()
+
+if __name__ == "__main__":
+    register()
